requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==5.0.6
django-wkhtmltopdf==3.4.0
firebase==4.0.1
fonttools==4.53.0
html5lib==1.1
idna==3.7
pillow==10.4.0
pycparser==2.22
pydyf==0.10.0
PyPDF2==3.0.1
pyphen==0.15.0
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
## The following requirements were added by pip freeze:
CacheControl==0.14.0
cachetools==5.3.3
cryptography==42.0.8
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
msgpack==1.0.8
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-magic==0.4.27
rsa==4.9
uritemplate==4.1.1